# swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "OAuth2 / OIDC authentication reverse proxy. Supports username/password login, OIDC provider federation, session management, and admin CRUD for users, groups, providers, and RBAC.",
        "title": "EzAuth API",
        "termsOfService": "https://github.com/flipcloud-ai/ezauth",
        "contact": {
            "name": "EzAuth Maintainers",
            "url": "https://github.com/flipcloud-ai/ezauth"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/ezauth/provider/": {
            "get": {
                "description": "Returns all OIDC/OAuth2 provider configurations from the database.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Provider Management"
                ],
                "summary": "List providers",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Max results (default 100)",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Offset for pagination (default 0)",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Provider list",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.ProviderDB"
                            }
                        }
                    },
                    "404": {
                        "description": "Database not available",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Adds a new OIDC/OAuth2 provider configuration to the database.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Provider Management"
                ],
                "summary": "Create a provider",
                "parameters": [
                    {
                        "description": "Provider configuration",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.ProviderDB"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Provider created"
                    },
                    "400": {
                        "description": "Invalid request body",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "Conflicts with a static provider",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/ezauth/provider/{name}": {
            "get": {
                "description": "Returns the OIDC/OAuth2 provider configuration.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Provider Management"
                ],
                "summary": "Get a provider by name",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Provider name",
                        "name": "name",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Provider configuration",
                        "schema": {
                            "$ref": "#/definitions/config.ProviderConfig"
                        }
                    },
                    "404": {
                        "description": "Provider not found",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            },
            "put": {
                "description": "Updates an existing OIDC/OAuth2 provider configuration in the database. The provider name is fixed and cannot be changed via this API.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Provider Management"
                ],
                "summary": "Update a provider",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Provider name",
                        "name": "name",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Updated provider fields",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.UpdateProviderRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Provider updated"
                    },
                    "400": {
                        "description": "Invalid request body",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Provider not found",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "Conflicts with a static provider",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "description": "Removes an OIDC/OAuth2 provider from the database and clears it from the cache.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Provider Management"
                ],
                "summary": "Delete a provider",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Provider name to delete",
                        "name": "name",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Provider deleted"
                    },
                    "404": {
                        "description": "Provider not found",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "dto.ErrorResponse": {
            "description": "Unified error response with numeric code and human-readable message.",
            "type": "object",
            "properties": {
                "authenticated": {
                    "type": "boolean"
                },
                "code": {
                    "type": "integer"
                },
                "error": {
                    "type": "string"
                }
            }
        },
...
        "models.LoginParameters": {
            "type": "object",
            "additionalProperties": {
                "type": "array",
                "items": {
                    "type": "string"
                }
            }
        },
        "models.ProviderDB": {
            "type": "object",
            "properties": {
                "admin_group": {
                    "type": "string"
                },
                "allowed_groups": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "authorization_endpoint": {
                    "$ref": "#/definitions/datatypes.URL"
                },
                "claims_from_profile": {
                    "type": "boolean"
                },
                "client_id": {
                    "type": "string"
                },
                "client_secret": {
                    "type": "string"
                },
                "code_challenge_methods_supported": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "created_at": {
                    "type": "string"
                },
                "device_auth_url": {
                    "$ref": "#/definitions/datatypes.URL"
                },
                "enabled": {
                    "type": "boolean"
                },
                "id_token_signing_alg_values_supported": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "issuer": {
                    "$ref": "#/definitions/datatypes.URL"
                },
                "jwks_uri": {
                    "$ref": "#/definitions/datatypes.URL"
                },
                "login_parameters": {
                    "$ref": "#/definitions/models.LoginParameters"
                },
                "protected_resource": {
                    "$ref": "#/definitions/datatypes.URL"
                },
                "provider_name": {
                    "type": "string"
                },
                "redirect_url": {
                    "$ref": "#/definitions/datatypes.URL"
                },
                "revocation_endpoint": {
                    "$ref": "#/definitions/datatypes.URL"
                },
                "scope": {
                    "type": "string"
                },
                "skip_nonce": {
                    "type": "boolean"
                },
                "token_endpoint": {
                    "$ref": "#/definitions/datatypes.URL"
                },
                "type": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "user_claim": {
                    "type": "string"
                },
                "userinfo_endpoint": {
                    "$ref": "#/definitions/datatypes.URL"
                },
                "validate_url": {
                    "$ref": "#/definitions/datatypes.URL"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "JWT bearer token obtained from a successful login. Prefix with \"Bearer \".",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        },
        "CookieAuth": {
            "description": "Session cookie set after login. Used by browser-based clients.",
            "type": "apiKey",
            "name": "_xw_session",
            "in": "cookie"
        }
    }
}